requirements.txt
asgiref==3.5.2
Django==4.0.5
django-extensions==3.2.0
django-filter==22.1
django-widget-tweaks==1.4.12
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
